Finish UI and logics (except make order)

+++ app/src/main/java/com/example/evaware/data/model/SavedModel.java
@@ -7,10 +7,10 @@ public class SavedModel {
     public DocumentReference productRef;
     public DocumentReference wishListRef;
     public String imageUrl;
-    public Float price;
+    public Double price;
     public String desc;
 
-    public SavedModel(DocumentReference wishListRef, DocumentReference productRef, String imageUrl, Float price, String desc) {
+    public SavedModel(DocumentReference wishListRef, DocumentReference productRef, String imageUrl, Double price, String desc) {
         this.wishListRef = wishListRef;
         this.productRef = productRef;
         this.imageUrl = imageUrl;
